Add basic CSV import test
